2016年 4月 5日 星期二 10时04分02秒 CST
选择诊所相关代码

+++ patient/Section/Info/InfoPerson/Model/ExpertClinicData.h
@@ -14,7 +14,7 @@
 @property (strong,nonatomic)NSString *outpat_name;
 @property (strong,nonatomic)NSString *address;
 @property (strong,nonatomic)NSString *commenResult;
-@property (assign,nonatomic)float money;
-@property (assign,nonatomic)float juli;
+@property (assign,nonatomic)double money;
+@property (assign,nonatomic)double juli;
 
 @end
